♻️ View 병합
